+++ src/wp_post.rs
@@ -7,8 +7,8 @@ pub mod post_status;
 #[derive(Debug)]
 #[allow(non_camel_case_types, non_snake_case)]
 pub struct WP_Post {
-    pub ID: u32,
-    pub post_author: u32,
+    pub ID: u64,
+    pub post_author: u64,
     pub post_date: Date,
     pub post_date_gmt: Date,
     pub post_content: String,
@@ -24,10 +24,10 @@ pub struct WP_Post {
     pub post_modified: Date,
     pub post_modified_gmt: Date,
     pub post_content_filtered: String,
-    pub post_parent: u32,
+    pub post_parent: u64,
     pub guid: String,
-    pub menu_order: u32,
+    pub menu_order: u64,
     pub post_type: String,
     pub post_mime_type: String,
-    pub comment_count: u32,
+    pub comment_count: u64,
 }
